+++ plugins/frontend/backstage-plugin-github-insights/src/apis/GithubClient.ts
@@ -15,8 +15,10 @@
  */
 
 import { GithubApi } from './GithubApi';
-import { ErrorApi, OAuthApi } from '@backstage/core-plugin-api';
+import { OAuthApi } from '@backstage/core-plugin-api';
 import { Octokit } from '@octokit/rest';
+import parseGitUrl from 'git-url-parse';
+import { MarkdownContentProps } from '../components/Widgets/MarkdownContent/types';
 
 const mimeTypeMap: Record<string, string> = {
   svg: 'image/svg+xml',
@@ -41,20 +43,12 @@ const defaultBaseUrl = 'https://api.github.com';
 
 export class GithubClient implements GithubApi {
   private githubAuthApi: OAuthApi;
-  private errorApi: ErrorApi;
 
-  constructor(deps: { githubAuthApi: OAuthApi; errorApi: ErrorApi }) {
+  constructor(deps: { githubAuthApi: OAuthApi }) {
     this.githubAuthApi = deps.githubAuthApi;
-    this.errorApi = deps.errorApi;
   }
 
-  async getContent(props: {
-    baseUrl?: string;
-    owner: string;
-    repo: string;
-    branch?: string;
-    path?: string;
-  }): Promise<{
+  async getContent(props: MarkdownContentProps): Promise<{
     content: string;
     media: Record<string, string>;
     links: Record<string, string>;
@@ -62,6 +56,7 @@ export class GithubClient implements GithubApi {
     const { path, repo, owner, branch, baseUrl = defaultBaseUrl } = props;
     const token = await this.githubAuthApi.getAccessToken();
     const octokit = new Octokit({ auth: token, baseUrl });
+
     let query = 'readme';
     if (path) {
       query = `contents/${path}`;
@@ -80,31 +75,56 @@ export class GithubClient implements GithubApi {
 
     const mediaLinks = [
       ...content.matchAll(
-        /\[([^\[\]]*)\]\((.*?)(\.png|\.jpg|\.jpeg|\.gif|\.webp|\.svg)(.*)\)/gim,
+        /!\[([^\]]*)\]\(([^)]+\.(?:png|jpg|jpeg|gif|webp|svg))\)/gi,
       ),
     ].map(match => [match[2], match[3]].join(''));
 
     const media: Record<string, string> = {};
 
-    for (const href of mediaLinks) {
-      const mimeType = mimeTypeLookup(href);
-      const url = new URL(href, response.data.url);
+    const { ref, resource, protocol } = parseGitUrl(response.data.url);
 
-      if (mimeType && url.host.includes('github.com')) {
-        const requestPath = url.pathname.replace(
-          new RegExp(`/${owner}/${repo}/blob/(main|master)`),
-          `/repos/${owner}/${repo}/contents`,
-        );
+    for (const mediaLink of mediaLinks) {
+      const mimeType = mimeTypeLookup(mediaLink);
+      if (!mimeType) {
+        continue;
+      }
+
+      const getUrl = () => {
+        const linkLowerCased = mediaLink.toLocaleLowerCase('en-US');
+        if (!linkLowerCased.startsWith('http')) {
+          return new URL(mediaLink, response.data.url);
+        }
+        if (linkLowerCased.includes('github')) {
+          const {
+            owner: ownerLink,
+            name: repoLink,
+            ref: refLink,
+            filepath,
+          } = parseGitUrl(mediaLink);
+          if (filepath) {
+            return `/repos/${ownerLink}/${repoLink}/contents/${filepath}${
+              refLink && `?ref=${refLink}`
+            }`;
+          }
+        }
+        return undefined;
+      };
+
+      const url = getUrl();
+      if (url) {
         try {
-          const contentResponse = await octokit.request(`GET ${requestPath}`);
+          const contentResponse = await octokit.request(`GET ${url}`);
           media[
-            href
+            mediaLink
           ] = `data:${mimeType};base64,${contentResponse.data.content.replaceAll(
             '\n',
             '',
           )}`;
         } catch (e: any) {
-          this.errorApi.post(e);
+          /* eslint no-console: ["error", { allow: ["warn"] }] */
+          console.warn(
+            `There was a problem loading the image content at ${url} via the GitHub API due to error: ${e.message}`,
+          );
         }
       }
     }
@@ -114,10 +134,14 @@ export class GithubClient implements GithubApi {
     ].map(match => [match[2], match[3]].join(''));
 
     const links: Record<string, string> = {};
+
+    const loadFromBranch =
+      branch || ref || getRepositoryDefaultBranch(response.data.url);
+
     for (const markdownLink of markdownLinks) {
-      links[markdownLink] = `https://github.com/${owner}/${repo}/blob/${
-        branch || getRepositoryDefaultBranch(response.data.url)
-      }/${markdownLink}`;
+      links[
+        markdownLink
+      ] = `${protocol}://${resource}/${owner}/${repo}/blob/${loadFromBranch}/${markdownLink}`;
     }
     return { content, media, links };
   }